Lint and separate test files
